+++ test/accessibility/sitemap-1.spec.js
@@ -3,7 +3,6 @@
 // over the first 1/4 of urls in the sitemap, while `sitemap-2-4.spec.js`
 // runs over the others. Crude, but this enables nightwatch to parallelize these.
 
-const Timeouts = require('../e2e/timeouts.js');
 const SitemapHelpers = require('../e2e/sitemap-helpers');
 
 module.exports = {
@@ -12,18 +11,7 @@ module.exports = {
     SitemapHelpers.sitemapURLs((urls, only508List) => {
       const mark = Math.ceil(urls.length / 4);
       const segment = urls.splice(0, mark);
-
-      segment.forEach(url => {
-        const only508 = only508List.filter(path => url.endsWith(path)).length > 0;
-        client
-          .perform(() => { console.log(url); }) // eslint-disable-line no-console
-          .url(url)
-          .waitForElementVisible('body', Timeouts.normal)
-          .axeCheck('document', only508 ?
-                    { scope: url, rules: ['section508'] } :
-                    { scope: url });
-      });
-
+      SitemapHelpers.runTests(client, segment, only508List);
       client.end();
     });
   }
