+++ test/accessibility/sitemap-3.spec.js
@@ -1,21 +1,12 @@
-const Timeouts = require('../e2e/timeouts.js');
 const SitemapHelpers = require('../e2e/sitemap-helpers');
 
 module.exports = {
   'sitemap 3/4': (client) => {
     client.timeoutsAsyncScript(1000);
-    SitemapHelpers.sitemapURLs(urls => {
+    SitemapHelpers.sitemapURLs((urls, only508List) => {
       const mark = Math.ceil(urls.length / 4);
       const segment = urls.splice(mark * 2, mark);
-
-      segment.forEach(url => {
-        client
-          .perform(() => { console.log(url); }) // eslint-disable-line no-console
-          .url(url)
-          .waitForElementVisible('body', Timeouts.normal)
-          .axeCheck('document', { scope: url });
-      });
-
+      SitemapHelpers.runTests(client, segment, only508List);
       client.end();
     });
   }
