+++ test/e2e/sitemap-helpers.js
@@ -1,6 +1,7 @@
 const libxmljs = require('libxmljs');
 const fetch = require('node-fetch');
 const E2eHelpers = require('./e2e-helpers');
+const Timeouts = require('./e2e/timeouts.js');
 
 const SITEMAP_URL = `${E2eHelpers.baseUrl}/sitemap.xml`;
 const SITEMAP_LOC_NS = 'http://www.sitemaps.org/schemas/sitemap/0.9';
@@ -13,9 +14,10 @@ function sitemapURLs(callback) {
     }).then((body) => {
       const doc = libxmljs.parseXml(body);
 
-      const urls = doc.find('//xmlns:loc', SITEMAP_LOC_NS)
-                      .map(n => n.text().replace(BUILD_BASE_URL, E2eHelpers.baseUrl))
-                      .filter(url => !(url.endsWith('auth/login/callback/')));
+      const urls = doc
+        .find('//xmlns:loc', SITEMAP_LOC_NS)
+        .map(n => n.text().replace(BUILD_BASE_URL, E2eHelpers.baseUrl))
+        .filter(url => !(url.endsWith('auth/login/callback/')));
 
       // Whitelist of URLs to only test against the 'section508' rule set and not
       // the stricter 'wcag2a' rule set. For each URL added to this list, please
@@ -32,5 +34,17 @@ function sitemapURLs(callback) {
     }).catch(callback);
 }
 
+function runTests(client, segment, only508List) {
+  segment.forEach(url => {
+    const only508 = only508List.filter(path => url.endsWith(path)).length > 0;
+    client
+      .perform(() => { console.log(url); }) // eslint-disable-line no-console
+      .url(url)
+      .waitForElementVisible('body', Timeouts.normal)
+      .axeCheck('document', only508 ?
+                { scope: url, rules: ['section508'] } :
+                { scope: url });
+  });
+}
 
-module.exports = { sitemapURLs };
+module.exports = { runTests, sitemapURLs };
